Write org variables into own R environment
* lisp/ob-R.el (org-babel-expand-body:R): Create empty environment
  called `.org-variables_' in R before adding variables and afterwards lock it and
  add it to the R search path. Save `.org-variables_' environment
  after variables are stored and locked into file of name
  `orgVARIABLES.RData'. Add beautifications to code submitted to R.
  (org-babel-variable-assignments:R): Assign variables into own
  `.org-variables_' environment in R instead of .GlobalEnv

These patch implements the writing of org variables into a separate R
environment and attaches it to the search path.  For the usage of these
variables, nothing changes in R, but:

1) The org variables are now grouped and can be seen via `ls(org)' in
R and are not shown anymore in the .GlobalEnv when using `ls()'
2) As the environment `org' and all bindings are locked, the variables
can not be accidentally deleted. They can be overwritten, but they can
be restored by simply deleting the variable in R or by using
`org$VARIABLE' instead of `VARIABLE'
3) All variables can be saved by simply calling `save(org, FILENAME'
in R which makes it possible to store all variable definitions for
tangling in one file.