Make transfer of values from R type aware
* lisp/ob-R.el (org-babel-R-assign-elisp): Added different cases for
transfer of integer, float, string and other variables to R so that
now integer values are transferred as integers (L) and stored in R as
such.  This change is backward compatible as integer values are
numerical values in R.
Moved definition of R functions for transfer of tables into defconst
ob-R-transfer-variable-table-with-header and
ob-R-transfer-variable-table-without-header.